Make Flag Sendable where its Value is Sendable
Previously, Flag was not Sendable. We can make it as such fairly easily if it’s carrying a Sendable Value (like a Bool or Int).

However, that requires the Decorator class to likewise be Sendable. Since it’s a class, we must manage the thread safety ourselves. We can use an NSLock for that, which works, but it’s perhaps a little manual and prone to developer-error — and perhaps to deadlock.

That also requires making the Decorator class final.

(Another option to that is to make it an actor, however that introduces a bit histrionics and drama of full async/await adoption — there may be other options too.)